LayoutTests/pointer-lock/lock-already-locked.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<div>
  <button onclick="doNextStep('manual');">doNextStep for manual testing</button>
  <div id="target1"></div>
  <div id="target2"></div>
</div>
<script>
    description("Test calling lock when already in a locked state.")
    window.jsTestIsAsync = true;

    targetdiv1 = document.getElementById("target1");
    targetdiv2 = document.getElementById("target2");

    // Expect change event only for all transitions below.
    expectedTargetToBeLockedString = "";
    document.onwebkitpointerlockchange = function ()
        {
            testPassed("document.onwebkitpointerlockchange event received.");
            shouldBe("document.webkitPointerLockElement", expectedTargetToBeLockedString);
            doNextStep();
        };
    document.onwebkitpointerlockerror =
        function () { testFailed("document.onwebkitpointerlockerror event received."); finishJSTest(); };

    var expectTarget1Unlock = false;
    targetdiv1.addEventListener("webkitpointerlocklost",
        function () { shouldBe("expectTarget1Unlock", "true"); });

    var expectTarget2Unlock = false;
    targetdiv2.addEventListener("webkitpointerlocklost",
        function () { shouldBe("expectTarget2Unlock", "true"); });

    currentStep = 0;
    function doNextStep(manual)
    {
        if (!window.layoutTestController && !manual)
            return;
        if (currentStep < todo.length)
            setTimeout(function () { todo[currentStep++](); }, 0);
        else if (currentStep++ == todo.length)
            setTimeout(function () { finishJSTest(); }, 0);
    }
    todo = [
        function () {
            shouldBe("document.webkitPointerLockElement", "null");
            debug("     Locking targetdiv1.")
            targetdiv1.webkitRequestPointerLock();
            expectedTargetToBeLockedString = "targetdiv1";
            // doNextStep() called by onwebkitpointerlockchange handler.
        },
        function () {
            debug("     Locking targetdiv1, again.")
            targetdiv1.webkitRequestPointerLock();
            expectedTargetToBeLockedString = "targetdiv1";
            // doNextStep() called by onwebkitpointerlockchange handler.
        },
        function () {
            debug("     Locking targetdiv2.")
            expectTarget1Unlock = true;
            targetdiv2.webkitRequestPointerLock();
            expectedTargetToBeLockedString = "targetdiv2";
            // doNextStep() called by onwebkitpointerlockchange handler.
        },
        function () {
            expectTarget1Unlock = false;
            expectTarget2Unlock = true;
            navigator.webkitPointer.lock(targetdiv1,
                function () {
                    testPassed("Lock target1 succeeded.");
                    shouldBe("navigator.webkitPointer.isLocked", "true");
                    // doNextStep() called by onwebkitpointerlockchange handler.
                },
                function () {
                    testFailed("Lock failed.");
                    finishJSTest();
                });
            expectedTargetToBeLockedString = "targetdiv1";
        },
        function () {
            expectTarget2Unlock = false;
            navigator.webkitPointer.lock(targetdiv1,
                function () {
                    testPassed("Lock target1 succeeded again.");
                    shouldBe("navigator.webkitPointer.isLocked", "true");
                    // doNextStep() called by onwebkitpointerlockchange handler.
                },
                function () {
                    testFailed("Lock failed.");
                    finishJSTest();
                });
            expectedTargetToBeLockedString = "targetdiv1";
        },
        function () {
            expectTarget1Unlock = true;
            navigator.webkitPointer.lock(targetdiv2,
                function () {
                    testPassed("Lock target2 succeeded.");
                    shouldBe("navigator.webkitPointer.isLocked", "true");
                    // doNextStep() called by onwebkitpointerlockchange handler.
                },
                function () {
                    testFailed("Lock failed.");
                    finishJSTest();
                });
            expectedTargetToBeLockedString = "targetdiv2";
        },
    ];
    doNextStep();
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
