LayoutTests/pointer-lock/lock-already-locked-expected.txt

Test calling lock when already in a locked state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.webkitPointerLockElement is null
     Locking targetdiv1.
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
     Locking targetdiv1, again.
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
     Locking targetdiv2.
PASS expectTarget1Unlock is true
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv2
PASS expectTarget2Unlock is true
PASS Lock target1 succeeded.
PASS navigator.webkitPointer.isLocked is true
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
PASS Lock target1 succeeded again.
PASS navigator.webkitPointer.isLocked is true
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
PASS expectTarget1Unlock is true
PASS Lock target2 succeeded.
PASS navigator.webkitPointer.isLocked is true
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv2
PASS successfullyParsed is true

TEST COMPLETE
doNextStep for manual testing
